move new examples
