+++ frontend/docs/pages/home/run-with-results.mdx
@@ -15,7 +15,7 @@ export const SimpleGo = {
 };
 
 export const FanoutPy = {
-  path: "examples/fanout/simple-fanout.py",
+  path: "examples/child/simple-fanout.py",
 };
 
 export const ParentTaskGo = {
