+++ sdks/python/examples/child/simple-fanout.py
@@ -1,6 +1,7 @@
 from datetime import timedelta
+from typing import Any
 
-from examples.simple.worker import SimpleInput, step1
+from examples.child.worker import SimpleInput, child_task
 from hatchet_sdk.context.context import Context
 from hatchet_sdk.hatchet import Hatchet
 from hatchet_sdk.runnables.types import EmptyModel
@@ -9,17 +10,12 @@ hatchet = Hatchet(debug=True)
 
 
 # ❓ Running a Task from within a Task
-parent_wf = hatchet.task(name="parent_task")
-
-
-@parent_wf.task(execution_timeout=timedelta(minutes=5))
+@hatchet.task(execution_timeout=timedelta(minutes=5))
 async def spawn(input: EmptyModel, ctx: Context) -> dict[str, Any]:
     # Simply run the task with the input we received
-    result = await step1.aio_run(
+    result = await child_task.aio_run(
         input=SimpleInput(message="Hello, World!"),
     )
 
     return {"results": result}
-
-
 # ‼️
