+++ sdks/python/examples/child/worker.py
@@ -15,15 +15,17 @@ class SimpleInput(EmptyModel):
 class SimpleOutput(BaseModel):
     transformed_message: str
 
+child_task = hatchet.workflow(name="SimpleWorkflow", input=SimpleInput)
 
-@hatchet.task(name="SimpleWorkflow")
+
+@child_task.task(name="step1")
 def step1(input: SimpleInput, ctx: Context) -> SimpleOutput:
     print("executed step1: ", input.message)
     return SimpleOutput(transformed_message=input.message.upper())
 
 
 def main() -> None:
-    worker = hatchet.worker("test-worker", slots=1, workflows=[step1])
+    worker = hatchet.worker("test-worker", slots=1, workflows=[child_task])
     worker.start()
